Moved ExtractKeyedValues to pipeline.util.
