+++ pipeline/util/__init__.py
@@ -4,6 +4,7 @@ from collections import defaultdict
 
 import settings
 import pipeline.io.arches
+from bonobo.config import Configurable, Option
 from cromulent.model import factory
 
 def identity(d):
@@ -114,3 +115,28 @@ class CromObjectMerger:
 		for v in unidentified:
 			setattr(obj, p, v)
 
+class ExtractKeyedValues(Configurable):
+	'''
+	Given a `dict` representing an some object, extract an array of `dict` values from
+	the `key` member. To each of the extracted dictionaries, add a 'parent_data' key with
+	the value of the original dictionary. Yield each extracted dictionary.
+	'''
+	key = Option(str, required=True)
+	include_parent = Option(bool, default=True)
+
+	def __init__(self, *v, **kw):
+		'''
+		Sets the __name__ property to include the relevant options so that when the
+		bonobo graph is serialized as a GraphViz document, different objects can be
+		visually differentiated.
+		'''
+		super().__init__(*v, **kw)
+		self.__name__ = f'{type(self).__name__} ({self.key})'
+
+	def __call__(self, data):
+		for a in data.get(self.key, []):
+			child = {k: v for k, v in a.items()}
+			child.update({
+				'parent_data': data,
+			})
+			yield child
